chore: Move VideoConfigService tests near inside its app
